# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto API
  description: |-
    This OpenAPI document describes the functionalities of Wasaphoto, a social media platform. The website has all the basic features such as banning/unbanning other users, liking/unliking photos, upload/delete photos and commenting/uncommenting on photos. 
    
    This API supports Cross-Origin Resource Sharing (CORS), allowing it to interact with other domains in this API for data reasons to increase the website's usability.
    
    This API was created in 2023/2024.
    
  version: "10"
paths:
...
  /username/{username}/follow/{followUsername}:
  
    get:
      tags: ["follow"]
      summary: check whether the user is following another user
      description: |-
        This endpoint checks whether a user is already following the user
      security:
        - bearerAuth: []
      operationId: checkFollowUser
      parameters:
        - in: path
          name: username
          required: true
          description: username of the current user.
          schema:
            type: string
        - in: path
          name: followUsername
          required: true
          description: check whether the the current user is following the other user with followUsername.
          schema:
              $ref: "#/components/schemas/Followers"
      responses:
        '200':
          description: check whether the user is following another user 
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFollowing:
                    type: boolean 
                    description: isFollowing is true if the user is following another user. Otherwise its false, if they are not already following the other user
                    example: true
        '400':
          description: UserID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "userID not found"
                  
          
  
    post:
      tags: ["follow"]
      summary: Follow a user
      description: |-
        This endpoint allows a user to follow or unfollow another user by their userID.
        If the user is already following the specified user, they will unfollow them, and if not, they will follow them.
      security:
        - bearerAuth: []
      operationId: followUser
      parameters:
        - in: path
          name: username
          required: true
          description: username of the current user
          schema:
            type: string
        - in: path
          name: followUsername
          required: true
          description: username of the person the account owner wants to follow
          schema:
            type: string
        
      
      responses:
        '200':
          description: User followed 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Follow"
        '400':
          description: UserID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "userID not found"
                  
          
          
    delete:
        tags: ["follow"]
        summary: Unfollow a user
        description: |-
          This endpoint allows a user to unfollow another user by their userID. 
        security:
          - bearerAuth: []
        operationId: unfollowUser
        parameters:
          - in: path
            name: username
            required: true
            description: current user's username
            schema:
              type: string
          - in: path
            name: followUsername
            required: true
            description: username of the person the account owner wants to unfollow
            schema:
              type: string
        responses:
          '200':
            description: User unfollowed 
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Follow"
          '400':
            description: UserID not found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: "Username not found"
  /username/{username}/ban/{banUsername}:
    get:
      tags: ["ban"]
      summary: Ban user
      description: |-
        This endpoint checks if the user is already banned
      security:
        - bearerAuth: []
      operationId: checkBanUser
      parameters:
        - in: path
          name: username
          required: true
          description: current user's username
          schema:
            $ref: "#/components/schemas/Username"
        - in: path
          name: banUsername
          required: true
          description: the username of a user the current user wants to ban
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  isBan:
                    type: boolean
                    description: isBan is either true or false. Its true if the current user has banned another user. False otherwise.
                    example: true
        '400':
          description: UserID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "UserID not found"
    post:
      tags: ["ban"]
      summary: Ban user
      description: |-
        This endpoint allows a user to ban another user by their userID.
        If the user is already banned, they will unban them, otherwise, they will ban them.
      security:
        - bearerAuth: []
      operationId: banUser
      parameters:
        - in: path
          name: username
          required: true
          description: current user's username
          schema:
            $ref: "#/components/schemas/Username"
        - in: path
          name: banUsername
          required: true
          description: the username of a user the current user wants to ban
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ban"
        '400':
          description: UserID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "UserID not found"
    delete:
        tags: ["ban"]
        summary: Unban a user
        description: |-
          This endpoint allows a user to unban another user by their userID. 
        security:
          - bearerAuth: []
        operationId: unbanUser
        parameters:
          - in: path
            name: username
            required: true
            description: current username
            schema:
              $ref: "#/components/schemas/Username"
          
          - in: path
            name: banUsername
            required: true
            description: username of the person the current user wants to unban
            schema:
              $ref: "#/components/schemas/Username"
        responses:
          '200':
            description: User unbanned successfully
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Ban"
            
          '400':
            description: UserID not found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: "UserID not found"
...
  /user/{userID}/photo/{photoID}/like:
    post:
      tags: ["like"]
      summary: Like a Photo
      description: |-
        This endpoint allows a user to like a photo. If the user has already liked the photo, it will unlike the photo. The photoID is a string that uniquely identifies each photo.
      security:
        - bearerAuth: []
      operationId: likePhoto
      parameters:
        - in: path
          name: photoID
          required: true
          description: adding a like to a collection of likes
          schema:
            type: string
            example: "picture1"
        - in: path
          name: userID
          required: true
          description: identification of the user
          schema:
            type: string
      responses:
        '200':
          description: Photo is liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Like"

    delete:
      tags: ["like"]
      summary: Unlike a photo
      description: |-
          This allows a user to unlike a photo.
      security:
        - bearerAuth: []
      operationId: unlikePhoto
      parameters:
        - in: path
          name: photoID
          required: true
          description: photoID to unlike
          schema:
            type: string
            example: "picture1"
        - in: path
          name: userID
          required: true
          description: ban user
          schema:
            type: string
      
      responses:
        '200':
          description: Photo is unliked 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Like"
  /photo/{photoID}/comment:
    get:
      tags: ["comment"]
      summary: get list of comments
      description: |-
        This endpoint allows a user to view the comments of a photo. 
      security:
        - bearerAuth: []
      operationId: getComments
      parameters:
        - in: path
          name: photoID
          required: true
          description: photoID is a unique identifier to retrieve the post in which we would like to post the comment.
          schema:
            type: string
            example: "picture1"
                
      responses:
        '200':
          description: User has commented on the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentList"  
  /user/{userID}/photo/{photoID}/comment:
    post:
      tags: ["comment"]
      summary: comment on a photo
      description: |-
        This endpoint allows a user to write a comment on a photo. 
      security:
        - bearerAuth: []
      operationId: commentPhoto
      parameters:
        - in: path
          name: photoID
          required: true
          description: photoID is a unique identifier to retrieve the post in which we would like to post the comment.
          schema:
            type: integer
            example: 12
        - in: path
          name: userID
          required: true
          description: userID of the person who comments on a photo
          schema:
            type: string
      requestBody:
        description: Here we comment on the photo according to the photoID given as parameter. 
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  example: "cool pic"
                
      responses:
        '200':
          description: User has commented on the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /uncomment/{commentID}:       
    delete:
      tags: ["comment"]
      summary: Uncomment on a photo
      description: |-
          This allows a user to uncomment on a photo using the commentID.
      security:
        - bearerAuth: []
      operationId: uncommentPhoto
      parameters:
        - in: path
          name: commentID
          required: true
          description: commentID is the identifier for a comment  
          schema:
            type: integer
            example: 1        
      responses:
        '200':
          description: comment deleted
          content:
            application/json:
              schema:
                type: object 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
  schemas:
    Username:
      type: object
      description: user's username
      properties:
        Username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
...
    Follow:
      type: object
      description: The userID of someone the current user would like to follow
      properties:
        followID:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^.*?$'
          description: The userID of the person you would like to follow
          example: "123-123-123"
    Followers:
      type: object
      description: the followers or followings of the user
      properties:
        followers:
          type: array
          minLength: 3
          maxLength: 100000
          pattern: '^.*?$'
          description: a list of all the followers of the user
          items:
            type: object
            description: the username of a user that follows the current user or that the current user follows. 
            properties:
              username:
                description: username.
                type: string
                pattern: '^.*?$'
                example: nata123
                minLength: 3
                maxLength: 16
    Ban:
      type: object
      description: A string that uniquely identifies a banned user in the banned users list.
      properties:
        banID:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^.*?$'
          description: The userID that the current user will ban 
          example: "123-123-123"
    Like:
      type: object
      description: a like on a photo
      properties:
        likeID: 
          type: string
          description: the like identifier.
          example: "123-123-123"
    Comment:
      type: object
      description: A comment
      properties:
        commentID:
          type: string
          description: A unique comment identifier
          pattern: '^.*?$'
          minLength: 2
          maxLength:  50
          example: "123-123-123"

        commentUser:
          type: string
          description: the username of the person who commented 
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "nata123" 
          
        text:
          type: string
          description: the comment as a string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 3000
          example: nice picture!


        photoID: 
          description:  the photo identifier.
          type: string
          pattern: '^.*?$'
          example: "123-123-123"
          minLength: 3
          maxLength: 50

        photoUser:
          type: string
          description: the username of the person who posted the photo
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "nata123" 
    CommentList:
      type: object
      description: list of comments on the photo
      properties:
        uploadedPhotos:
          type: array
          description: list of comments on a photo 
          minItems: 0
          maxItems: 9999999
          items:
            type: object
            properties:
              photoID: 
                description:  the photo identifier.
                type: string
                pattern: '^.*?$'
                example: "123-123-123"
                minLength: 3
                maxLength: 50
              comment:
                description: the comment posted on the photo
                type: string
                pattern: '^.*?$'
                example: nice picture!
                minLength: 3
                maxLength: 10000
              username:
                description: the username of the person who wrote the comment on the photo
                type: string
                pattern: '^.*?$'
                example: nata12 
                minLength: 3
                maxLength: 10000